refactor : 인터페이스를 사용부에 위치
