+++ src/main/java/roomescape/repository/DatabaseReservationTimeRepository.java
@@ -7,6 +7,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import roomescape.controller.ReservationTimeRepository;
 import roomescape.domain.ReservationTime;
 
 @Repository
